+++ internal/app/api/api.go
@@ -3,11 +3,11 @@ package api
 import (
 	"net/http"
 
-	memberhandler "dating/internal/app/api/handler/member"
 	userhandler "dating/internal/app/api/handler/user"
+	user "dating/internal/app/api/repositories"
+	userService "dating/internal/app/api/services/user"
+
 	"dating/internal/app/db"
-	"dating/internal/app/member"
-	"dating/internal/app/user"
 	"dating/internal/pkg/glog"
 	"dating/internal/pkg/health"
 	"dating/internal/pkg/middleware"
@@ -42,22 +42,16 @@ const (
 func Init(conns *InfraConns) (http.Handler, error) {
 	logger := glog.New()
 
-	var memberRepo member.Repository
-	var userRepo user.Repository
+	var userRepo userService.Repository
 	switch conns.Databases.Type {
 	case db.TypeMongoDB:
-		memberRepo = member.NewMongoRepository(conns.Databases.MongoDB)
 		userRepo = user.NewMongoRepository(conns.Databases.MongoDB)
 	default:
 		panic("database type not supported: " + conns.Databases.Type)
 	}
 
-	memberLogger := logger.WithField("package", "member")
-	memberSrv := member.NewService(memberRepo, memberLogger)
-	memberHandler := memberhandler.New(memberSrv, memberLogger)
-
 	userLogger := logger.WithField("package", "user")
-	userSrv := user.NewService(userRepo, userLogger)
+	userSrv := userService.NewService(userRepo, userLogger)
 	userHandler := userhandler.New(userSrv, userLogger)
 
 	routes := []route{
@@ -68,12 +62,6 @@ func Init(conns *InfraConns) (http.Handler, error) {
 			handler: health.Readiness().ServeHTTP,
 		},
 		// services
-		route{
-			path:    "/api/v1/member/{id:[a-z0-9-\\-]+}",
-			method:  get,
-			handler: memberHandler.Get,
-		},
-		//service sign up
 		route{
 			path:    "/signup",
 			method:  post,
