+++ internal/app/api/types/hobby.go
