+++ internal/app/api/types/user.go
@@ -1,6 +1,10 @@
 package types
 
-import "github.com/globalsign/mgo/bson"
+import (
+	"time"
+
+	"github.com/globalsign/mgo/bson"
+)
 
 type User struct {
 	ID       bson.ObjectId   `json:"_id" bson:"_id,omitempty"`
@@ -12,13 +16,20 @@ type User struct {
 	Country  string          `json:"country" bson:"country"`
 	Hobby    []Hobby         `json:"hobby" bson:"hobby"`
 	About    string          `json:"about" bson:"about"`
-	Like_id  []bson.ObjectId `json:"like_id" bson:"like_id"`
-	Math     []bson.ObjectId `json:"match" bson:"math"`
+	LikeID   []bson.ObjectId `json:"like_id" bson:"like_id"`
+	MatchID  []bson.ObjectId `json:"match_id" bson:"match_id"`
+	CreateAt time.Time       `json:"created_at" bson:"created_at"`
+	UpdateAt time.Time       `json:"updated_at" bson:"updated_at"`
 }
 type UserSignUp struct {
 	Name     string `json:"name"`
 	Email    string `json:"email"`
 	Password string `json:"password"`
+}
+type UserFieldInToken struct {
+	ID    bson.ObjectId `json:"_id" bson:"_id,omitempty"`
+	Name  string        `json:"name"`
+	Email string        `json:"email"`
 }
 type UserResponseSignUp struct {
 	Name  string `json:"name"`
