--- internal/app/types/member.go
@@ -1,7 +1,0 @@
-package types
-
-// member hold information of a member
-type Member struct {
-	ID   string `json:"id" bson:"id"`
-	Name string `json:"name"`
-}
